fix: restructure for deployment
