+++ .gitignore
@@ -1,0 +1,6 @@
+node_modules
+.next
+out
+.DS_Store
+public/sw.js
+public/workbox-*.js
