+++ src/app/layout.tsx
@@ -1,3 +1,6 @@
+import "./globals.css";
+import type { ReactNode } from "react";
+
 export const metadata = {
   title: "Horse Coat Color Calculator",
   description: "A glam UI to explore horse coat genetics",
@@ -5,7 +8,7 @@ export const metadata = {
   icons: { icon: "/icons/icon-192.png", apple: "/icons/icon-192.png" }
 };
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+export default function RootLayout({ children }: { children: ReactNode }) {
   return (
     <html lang="en">
       <body>{children}</body>
