+++ src/components/ui/card.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "./utils";
+import { cn } from "@/lib/utils";
 
 export function Card({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) {
   return <div className={cn("bg-[var(--card)]/80 backdrop-blur border border-white/10 rounded-2xl p-6", className)} {...props} />;
